+++ push_swap.h
@@ -6,7 +6,7 @@
 /*   By: ochoumou <ochoumou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/23 12:24:10 by ochoumou          #+#    #+#             */
-/*   Updated: 2022/04/05 02:30:46 by ochoumou         ###   ########.fr       */
+/*   Updated: 2022/04/06 00:15:43 by ochoumou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 
 # include <unistd.h>
 # include <stdlib.h>
+# include <stdio.h>
 
 # define INT_MIN -2147483648
 # define INT_MAX 2147483647
@@ -78,7 +79,7 @@ int		check_sorted_stack(t_stack *stack);
 int		push_el(t_stack *stack, int value);
 int		ft_isdigit(int c);
 int		ft_strlen(char *str);
-int		ft_atoi(char *str);
+long	ft_atoi(char *str);
 int		ft_strcmp(char *s1, char *s2);
 int		find_el(t_stack *stack, int value);
 int		stack_size(t_stack *stack);
